refactoring in the car
